moving functions around some more
